Removed extraneous extensions helpers
